+++ CHANGELOG.md
@@ -5,7 +5,8 @@ v2.2.0
 ------
 * Add a convenience synchronized_ method.
 * Add methods to Variable and Label to obtain the MethodMaker they belong to.
-* Workaround for dynamic constant bug isn't required with Java 19.
+* Check for string constants which are too large for the modified UTF-8 form.
+* Remove a workaround for a dynamic constant bug when using Java 19.
 
 v2.1.0 (2022-01-27)
 ------